fix(calendar): persist reverse-sync failure state, add DB test
MarkCalendarEventSyncFailed was failing with
`pq: inconsistent types deduced for parameter $2` whenever the worker
tried to record a failure. pq's parameter-type inference couldn't
reconcile `last_attempt_at = $2` (target bigint) with
`next_attempt_at = $2 + CASE ... END` (CASE branches inferred as int),
so the prepare itself errored out — retry_count never advanced,
last_error stayed empty, and the row sat in the queue being retried
every minute forever. Concretely visible as stale Google-Calendar
sync entries from May with retry_count=0.

Fix: cast `$2::bigint` in both usages so the deduction is unambiguous.

Also adds internal/db/calendar_event_sync_test.go — an integration
test gated by MAILSERVER_TEST_DSN that asserts retry_count is bumped,
last_error stored, and the first-retry backoff is 60s. Local run:

  ssh -fN -L 15432:10.0.0.2:5432 lucky@mail.letotam.ru
  MAILSERVER_TEST_DSN='host=127.0.0.1 port=15432 user=mailserver \
    password=mailserver dbname=mailserver sslmode=disable' \
    go test ./internal/db/

+++ internal/db/calendar_event_sync.go
@@ -83,12 +83,16 @@ func (db *DB) MarkCalendarEventSyncFailed(id int64, errMsg string) error {
 		errMsg = errMsg[:1024] + "…"
 	}
 	now := timeutil.Now()
+	// $2 is cast to bigint explicitly: pq's parameter-type inference can
+	// otherwise deduce different types for the same placeholder when it
+	// appears in two arithmetic contexts (assign vs. add-then-assign),
+	// failing the prepare with "inconsistent types deduced for parameter $2".
 	query := `
 		UPDATE calendar_event_sync_queue
 		SET retry_count = retry_count + 1,
 		    last_error = $1,
-		    last_attempt_at = $2,
-		    next_attempt_at = $2 + CASE
+		    last_attempt_at = $2::bigint,
+		    next_attempt_at = $2::bigint + CASE
 		        WHEN retry_count + 1 = 1 THEN 60000
 		        WHEN retry_count + 1 = 2 THEN 300000
 		        WHEN retry_count + 1 = 3 THEN 1800000
